complete directory refactor for pytest and main
