+++ website_analysis/tests/resources/cs417-one-lecture/integralDI2.png
