+++ website_analysis/tests/writer/test_excelWriter.py
@@ -1,22 +1,21 @@
 import pytest
-from pathlib import Path
+from pathlib import Path  
 
-from website_analysis.writers.json_writer import JsonWriter
-from website_analysis.website.site import Website
+from website_analysis.src.writers.excel_writer import ExcelWriter 
 
 
 @pytest.fixture
 def writer():
-    j = JsonWriter("tests/output/test")
-    yield j
+    e = ExcelWriter("tests/output/test")
+    yield e 
 
 
-class TestJsonWriter:
+class TestExcelWriter:
     def test_fullPath(self, writer):
-        assert writer.fullPath == Path(Path.cwd() / "tests/output/test.json")
-
-        writer.fullPath = "build/file"
-        assert writer.fullPath == Path(Path.cwd() / "build/file.json")
+        assert writer.fullPath == Path.cwd() / "tests/output/test.xlsx"
+        
+        writer.fullPath = "build/exlFile"
+        assert writer.fullPath == Path.cwd() / "build/exlFile.xlsx"
 
     def test_write(self, writer):
         writer.write()
